delete locust user without api endpoint & modify the locust tutorial
