--- dff/utils/locust_user/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-# flake8: noqa: F401
-
-from .locust_user import PipelineUser, IncorrectResponse
